Revert "[hiroi_9]バックエンドでマイグレーション適応"
