+++ .env.backend.example
@@ -4,4 +4,4 @@ DB_PASSWORD=your_rds_mysql_password
 DB_NAME=your_database_name
 DB_PORT=3306
 FLASK_SECRET_KEY=a_very_secret_key_for_flask_sessions
-FLASK_APP=backend.run
+FLASK_APP=app:create_app
