--- .github/pull_request_template.md
@@ -1,18 +1,0 @@
-<!-- I want to review in Japanese. -->
-## 概要
-- 
-
-## 修正内容の詳細
-
-## レビューで注意してみてほしい観点
-
-### レビューに関して
-レビューする際には、以下のprefix(接頭辞)を付けましょう。
-<!-- for GitHub Copilot review rule -->
-[must] → かならず変更してね  
-[imo] → 自分の意見だとこうだけど修正必須ではないよ(in my opinion)  
-[nits] → ささいな指摘(nitpick) 
-[ask] → 質問  
-[fyi] → 参考情報
-<!-- for GitHub Copilot review rule-->
-<!-- I want to review in Japanese. -->
