+++ .gitignore
@@ -1,11 +1,12 @@
 # Python
 __pycache__/
 *.pyc
+.env.backend
 
 # Node/React
 node_modules/
 build/
-.env*
+.env.frontend
 
 # Docker
 .dockerignore
