+++ backend/requirements.txt
@@ -2,5 +2,3 @@ Flask==2.3.3
 pymysql==1.0.2
 Flask-Cors==3.0.10
 Werkzeug==2.3.7
-Flask-Migrate==4.0.4
-Flask-SQLAlchemy==3.1.1
